根据Model调整Controller位置
