+++ src/main/java/com/gxh/controller/CommonController.java
@@ -1,24 +1,13 @@
 package com.gxh.controller;
 
-import com.baomidou.mybatisplus.core.conditions.query.LambdaQueryWrapper;
-import com.baomidou.mybatisplus.core.conditions.query.QueryWrapper;
-import com.gxh.entity.Category;
-import com.gxh.service.ICategoryService;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
-import org.springframework.ui.Model;
-import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 
 import javax.servlet.http.HttpSession;
-import java.util.List;
 
 @Controller
 @RequestMapping("/Common")
 public class CommonController {
-
-    @Autowired
-    ICategoryService categoryService;
 
     @RequestMapping("/login")
     public String userlogin(){
@@ -78,15 +67,6 @@ public class CommonController {
     @RequestMapping("/categoryUpdata")
     public String categoryUpdata(){
         return "category/categoryUpdata";
-    }
-
-    @RequestMapping("/categoryUpdata/{id}")
-    public String categoryUpdataById(Model model, @PathVariable Integer id){
-        LambdaQueryWrapper<Category> wrapper=new LambdaQueryWrapper<>();
-        wrapper.eq(Category::getId,id);
-        List<Category> list = categoryService.list(wrapper);
-        model.addAttribute("category",list);
-        return "category/categoryUpdataModel";
     }
 
     @RequestMapping("/productList")
