+++ src/main/resources/templates/category/categoryList.html
@@ -187,7 +187,7 @@
                             title: '编辑 - id:'+ data.id,
                             type: 2,
                             area: ['80%','80%'],
-                            content: '/BootStart/Common/categoryUpdata/'+data.id,//model方式获取修改页面数据
+                            content: '/BootStart/ModelData/categoryUpdata/'+data.id,//model方式获取修改页面数据
                             // content: '/BootStart/Common/categoryUpdata',
                             end:function (){tb_user.reload('user');}//重新加载表格
                         });break;
